+++ lib/google-finance/api.rb
@@ -1,2 +1,3 @@
 require_relative 'api/index'
 require_relative 'api/get_prices'
+require_relative 'api/historical'