+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ### 0.2.1 (Next)
 
+* [#5](https://github.com/dblock/google-finance-ruby-client/issues/5): Added support for interday historical price data - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.2.0 (2017/12/28)
 
-* [#1](https://github.com/dblock/google-finance-ruby-client/issues/1): Retrieve history of prices - [@dblock](https://github.com/dblock).
+* [#1](https://github.com/dblock/google-finance-ruby-client/issues/1): Added support for intraday historical price data - [@dblock](https://github.com/dblock).
 
 ### 0.1.0 (2017/12/02)